calculating payouts implemented

+++ controllers/user.js
@@ -2,6 +2,7 @@ import { validationResult } from "express-validator";
 import { Employee } from "../models/employee.js";
 import postAddEditEmployee from "../helpers/employeeAddEdit.js";
 import { Campaign } from "../models/campaign.js";
+import { ObjectId } from "mongodb";
 
 export const getHome = (req, res, next) => {
   console.log(req.user);
@@ -38,29 +39,43 @@ export const getAddEmployee = (req, res, next) => {
 };
 
 export const getEditEmployee = async (req, res, next) => {
-  const foundEmployee = await Employee.findById(req.params.employeeId);
-  if (foundEmployee.bossId.toString() !== req.user._id.toString()) {
-    const error = new Error("Forbidden operation");
-    error.httpStatusCode = 403;
+  try {
+    if (!ObjectId.isValid(req.params.employeeId)) {
+      throw new Error("Employee not found");
+    }
+    const foundEmployee = await Employee.findById(req.params.employeeId);
+    console.log(foundEmployee);
+    if (!foundEmployee) {
+      throw new Error("Employee not found");
+    }
+    if (foundEmployee.bossId.toString() !== req.user._id.toString()) {
+      throw new Error("Forbidden operation");
+    }
+    const { name, surname, hourlyRate, dailyHours, _id } = foundEmployee;
+    const oldInput = {
+      name,
+      surname,
+      hourlyRate,
+      dailyHours,
+      delegationAmount: foundEmployee.delegationAmount,
+      id: _id.toString(),
+    };
+    console.log(oldInput);
+    res.render("user/add-employee", {
+      pageTitle: "Edytuj pracownika",
+      errors: [],
+      oldInput,
+      isUserSigned: undefined,
+      edit: true,
+    });
+  } catch (error) {
+    if (error.message === "Employee not found") {
+      error.httpStatusCode = 404;
+    } else {
+      error.httpStatusCode = 403;
+    }
     return next(error);
   }
-  const { name, surname, hourlyRate, dailyHours, _id } = foundEmployee;
-  const oldInput = {
-    name,
-    surname,
-    hourlyRate,
-    dailyHours,
-    delegationAmount: foundEmployee.delegationAmount,
-    id: _id.toString(),
-  };
-  console.log(oldInput);
-  res.render("user/add-employee", {
-    pageTitle: "Edytuj pracownika",
-    errors: [],
-    oldInput,
-    isUserSigned: undefined,
-    edit: true,
-  });
 };
 
 export const postAddEmployee = async (req, res, next) => {
@@ -79,20 +94,30 @@ export const postEditEmployee = async (req, res, next) => {
   try {
     await postAddEditEmployee(req, res, next, errors, "Edit");
   } catch (error) {
-    error.message = "Server bug";
-    error.httpStatusCode = 500;
+    if (error.message === "Employee not found") {
+      error.httpStatusCode = 404;
+    } else {
+      error.message = "Server bug";
+      error.httpStatusCode = 500;
+    }
     return next(error);
   }
 };
 
 export const postDeleteEmployee = async (req, res, next) => {
-  const employee = await Employee.findById(req.params.employeeId);
-  if (employee.bossId.toString() !== req.user._id.toString()) {
-    const error = new Error("Forbidden operation");
-    error.httpStatusCode = 403;
-    return next(error);
-  }
   try {
+    if (!ObjectId.isValid(req.params.employeeId)) {
+      throw new Error("Employee not found");
+    }
+    const employee = await Employee.findById(req.params.employeeId);
+    if (!foundEmployee) {
+      throw new Error("Employee not found");
+    }
+    if (employee.bossId.toString() !== req.user._id.toString()) {
+      const error = new Error("Forbidden operation");
+      error.httpStatusCode = 403;
+      return next(error);
+    }
     await employee.deleteOne();
     res.redirect("/employees");
   } catch (error) {
@@ -158,28 +183,78 @@ export const postDeleteCampaign = async (req, res, next) => {
 };
 
 export const getCampaignDetails = async (req, res, next) => {
-  const presentCampaign = await Campaign.where({
-    ownerId: req.user._id,
-  })
-    .findOne()
-    .populate("employeesData.employeeId");
-  res.render("user/campaign-details", {
-    pageTitle: "Szczegóły kampanii",
-    employees: presentCampaign.employeesData,
-  });
+  try {
+    const presentCampaign = await Campaign.where({
+      ownerId: req.user._id,
+    })
+      .findOne()
+      .populate("employeesData.employeeId");
+    return res.render("user/campaign-details", {
+      pageTitle: "Szczegóły kampanii",
+      employees: presentCampaign.employeesData,
+    });
+  } catch (error) {
+    error.message = "Server bug";
+    error.httpStatusCode = 500;
+    return next(error);
+  }
 };
 
 export const getEmployeeDetails = async (req, res, next) => {
-  const presentCampaign = await Campaign.where({
-    ownerId: req.user._id,
-  })
-    .findOne()
-    .populate("employeesData.employeeId");
-  let employee;
-  presentCampaign.employeesData.forEach((emp) => {
-    if (emp.employeeId._id.toString() === req.params.employeeId.toString()) {
-      employee = { ...emp["_doc"] };
+  try {
+    if (!ObjectId.isValid(req.params.employeeId)) {
+      throw new Error("Employee not found");
     }
-  });
-  res.render("user/employee-details", { pageTitle: "Sczegóły pracownika" });
+    const presentCampaign = await Campaign.where({
+      ownerId: req.user._id,
+    })
+      .findOne()
+      .populate("employeesData.employeeId");
+    let employee;
+    const employeeIndex = presentCampaign.employeesData.findIndex((emp) => {
+      return emp.employeeId._id.toString() === req.params.employeeId.toString();
+    });
+    employee = { ...presentCampaign.employeesData[employeeIndex]["_doc"] };
+    if (!employee) {
+      throw new Error("Employee not found");
+    }
+    console.log(employee);
+    const campaignEndTime = presentCampaign.endtime;
+    return res.render("user/employee-details", {
+      pageTitle: "Sczegóły pracownika",
+      employee,
+      campaignEndTime,
+    });
+  } catch (error) {
+    if (error.message === "Employee not found") {
+      error.httpStatusCode = 404;
+    } else {
+      error.message = "Server bug";
+      error.httpStatusCode = 500;
+    }
+    return next(error);
+  }
+};
+
+export const postUpdateEmployeeWorkdays = async (req, res, next) => {
+  const selectedDays = req.body["dates"].split(", ");
+  console.log(selectedDays);
+  try {
+    const presentCampaign = await Campaign.where({
+      ownerId: req.user._id,
+    })
+      .findOne()
+      .populate("employeesData.employeeId");
+    let employee;
+    const employeeIndex = presentCampaign.employeesData.findIndex((emp) => {
+      return emp.employeeId._id.toString() === req.params.employeeId.toString();
+    });
+    presentCampaign.employeesData[employeeIndex].workdays = selectedDays;
+    await presentCampaign.save();
+    return res.redirect("/campaign");
+  } catch (error) {
+    error.message = "Server bug";
+    error.httpStatusCode = 500;
+    return next(error);
+  }
 };
